Various tweaks/cleanup
