+++ MyDataHelpsKit/API/ConnectExternalAccountResource.swift
@@ -1,0 +1,19 @@
+//
+//  ConnectExternalAccountResource.swift
+//  MyDataHelpsKit
+//
+//  Created by CareEvolution on 8/30/21.
+//
+
+import Foundation
+
+struct ConnectExternalAccountResource: ParticipantResource {
+    typealias ResponseType = URL
+    
+    let providerID: Int
+    let finalRedirectURL: URL
+
+    func urlRequest(session: ParticipantSession) throws -> URLRequest {
+        return session.authenticatedRequest(.POST, url: try session.client.endpoint(path: "api/v1/delegated/externalaccountproviders/\(providerID)/connect", queryItems: [.init(name: "finalRedirectPath", value: finalRedirectURL.absoluteString)]))
+    }
+}
